Added a toolbar to the main window with pause, play, and cycle buttons.
I've noticed a problem with shot collisions I can't figure out.  Occassionally, shots will seem to collide with nothing.

Extensive bug testing would be easier if I could figure out the selection code (that is, picking an object to examine the details of with the mouse).

+++ GUI/ViewerControls.cpp
@@ -102,7 +102,7 @@ unsigned int MainWindow::Selection(unsigned int MouseX, unsigned int MouseY)
     GLdouble height = canvas->getHeight();
     GLdouble aspect = height>0 ? width/height : 1.0;
 
-    glViewport(0,0,width,height);
+    glViewport(0,0,(int)width,(int)height);
 
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
